pyproject.toml
[project]
name = "sktk"
version = "0.1.0"
description = "Production-grade Semantic Kernel. Without the boilerplate."
license = {text = "GPL-3.0-only"}
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "SKTK Contributors"}]
keywords = ["semantic-kernel", "llm", "agents", "rag", "orchestration"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "semantic-kernel>=1.0.0,<2.0.0",
    "pydantic>=2.0",
    "packaging>=23.0",
]

[project.urls]
Documentation = "https://github.com/mangobanaani/sktk"
Repository = "https://github.com/mangobanaani/sktk"
Issues = "https://github.com/mangobanaani/sktk/issues"
Changelog = "https://github.com/mangobanaani/sktk/releases"

[project.optional-dependencies]
redis = ["redis>=5.0"]
chroma = ["chromadb>=0.4"]
pgvector = ["asyncpg>=0.29", "pgvector>=0.2"]
otel = [
    "opentelemetry-api>=1.24",
    "opentelemetry-sdk>=1.24",
    "opentelemetry-exporter-otlp>=1.24",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "mypy>=1.8",
    "ruff>=0.2",
    "bandit>=1.7",
]
openai = ["openai>=1.0"]
openapi = ["httpx>=0.27"]
mcp = ["mcp>=1.0"]
a2a = ["httpx>=0.27"]
rag-faiss = ["faiss-cpu>=1.8"]
rag-hnsw = ["hnswlib>=0.8"]
persistence = ["aiosqlite>=0.19,<1.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sktk"]

[tool.hatch.version]
path = "src/sktk/__init__.py"
pattern = '__version__ = "(?P<version>[^"]+)"'

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: tests that hit external APIs (requires ANTHROPIC_API_KEY)",
    "e2e: end-to-end workflow tests (requires ANTHROPIC_API_KEY)",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]
ignore = ["E501", "UP035", "SIM108"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["sktk"]

[tool.coverage.run]
source = ["sktk"]

[tool.coverage.report]
fail_under = 95
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@abstractmethod",
]

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]

[tool.mypy]
python_version = "3.11"
strict = true
